requirements.txt
numpy
pandas
protobuf
python-dotenv
google-generativeai
streamlit
streamlit-shadcn-ui
streamlit-chatbox
streamlit-aggrid
pyodbc
openai
faiss-cpu
sqlalchemy
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-core>=1.29.0